requirements.txt
envparse==0.2.0
flask==1.0.2
gunicorn==19.9.0
marshmallow==3.0.0rc4
Flask-Cors==3.0.7
requests==2.21.0
boto3==1.9.184
adapt-parser==0.3.3
ibm-watson==3.1.1